% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_timecci.R
\name{cal_timecci}
\alias{cal_timecci}
\title{Temporal Cell-Cell Interaction Analysis in Single-Cell RNA-seq Data}
\usage{
cal_timecci(
  seu.obj,
  Pseudotime.name = "Pseudotime",
  CellType.name = "CellType",
  discrete.numer = 20,
  min.cells = 10,
  CCI.database = CellChat::CellChatDB.human,
  interaction.name = "interaction_name_2"
)
}
\arguments{
\item{seu.obj}{A Seurat object that contains single-cell RNA-seq data. The object must include metadata columns for pseudotime and cell type.}

\item{Pseudotime.name}{The name of the metadata column in \code{seu.obj} that contains pseudotime information. Default is 'Pseudotime'.}

\item{CellType.name}{The name of the metadata column in \code{seu.obj} that contains cell type information. Default is 'CellType'.}

\item{discrete.numer}{The number of discrete time points to divide the pseudotime into. Default is 20.}

\item{min.cells}{The minimum number of cells required for a cell-cell interaction to be considered. Default is 10.}

\item{CCI.database}{The cell-cell interaction database to use. Default is \code{CellChat::CellChatDB.human}.}

\item{interaction.name}{The name of the column in the CellChat LRsig table that contains the interaction name. Default is "interaction_name_2".}
}
\value{
A list containing:
\item{cor.res}{A data frame with the correlation results, including p-values and correlation coefficients.}
\item{cal.res}{A data frame with the calculated ligand-receptor interaction probabilities over time.}
\item{cellchat}{The updated CellChat object after the analysis.}
}
\description{
This function performs temporal cell-cell interaction analysis in single-cell RNA-seq data using the CellChat framework.
It calculates correlations between ligand-receptor interaction probabilities and pseudotime across different cell states.
}
\details{
The function first removes any missing data from the Seurat object's metadata. It then divides the pseudotime into discrete intervals
and adds this information as new metadata to the Seurat object. The CellChat object is then created and used to compute communication
probabilities between cells based on their temporal states. Finally, correlations between ligand-receptor interaction probabilities
and pseudotime are calculated.
}
