% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LRdf}
\alias{LRdf}
\title{LRdf: Ligand-Receptor Interaction Database}
\format{
A data frame with columns:
\describe{
\item{interaction_name}{The name of the ligand-receptor interaction.}
\item{ligand.symbol}{The symbol of the ligand gene.}
\item{receptor.symbol}{The symbol of the receptor gene.}
}
}
\source{
CellChatDB.human$interaction
}
\usage{
LRdf
}
\description{
A data frame containing ligand-receptor interactions derived from the CellChat database.
This dataset includes the interaction name, ligand symbol, and receptor symbol.
}
\examples{
data(LRdf)
head(LRdf)
}
\keyword{datasets}
