% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalCCI1.R
\name{TemporalCCI1}
\alias{TemporalCCI1}
\title{Temporal Cell-Cell Interaction Analysis via Cell Alignment in Pseudotime}
\usage{
TemporalCCI1(
  seu.obj,
  CellType.name = "Cell.Type",
  Pseudotime.name = "Pseudotime",
  CCI.database = LRdf,
  source.celltype = "mCAF",
  target.celltype = "EpiT",
  window.size = 0.1,
  num.pts = 200,
  num.cores = 10,
  granger.cutoff = 0.01
)
}
\arguments{
\item{seu.obj}{A Seurat object containing single-cell RNA-seq data and metadata.}

\item{CellType.name}{A string specifying the name of the cell type annotation in the Seurat object metadata. Default is "Cell.Type".}

\item{Pseudotime.name}{A string specifying the name of the pseudotime annotation in the Seurat object metadata. Default is "Pseudotime".}

\item{CCI.database}{A data frame containing ligand-receptor interaction pairs with columns "interaction_name", "ligand.symbol", and "receptor.symbol".}

\item{source.celltype}{A string specifying the source cell type for the ligand in the CCI analysis. Default is "mCAF".}

\item{target.celltype}{A string specifying the target cell type for the receptor in the CCI analysis. Default is "EpiT".}

\item{window.size}{A numeric value defining the window size for trajectory alignment. Default is 0.1.}

\item{num.pts}{An integer specifying the number of points for trajectory interpolation. Default is 200.}

\item{num.cores}{An integer specifying the number of CPU cores to use for parallel processing. Default is 10.}

\item{granger.cutoff}{A numeric value for the p-value cutoff in Granger causality tests. Default is 0.01.}
}
\value{
A list containing:
\item{res}{A data frame with correlation results between ligand and receptor pairs, including Pearson and Spearman correlation coefficients and p-values.}
\item{source.align.matrix}{A data frame with the aligned and interpolated expression matrix of the source cell type.}
\item{target.align.matrix}{A data frame with the aligned and interpolated expression matrix of the target cell type.}
}
\description{
This function performs a temporal cell-cell interaction (CCI) analysis between two specified cell types over pseudotime.
It leverages ligand-receptor interaction data to assess the correlation between ligand expression in one cell type
and receptor expression in another across a defined pseudotime trajectory.
}
\details{
This function is designed to analyze temporal changes in cell-cell communication by evaluating ligand-receptor pair correlations
over pseudotime. It uses interpolation methods to align gene expression trajectories, enabling the comparison of expression patterns
between a source and a target cell type. The function filters out ligand-receptor pairs that are not present in the RNA assay of
the Seurat object. Correlation analyses are performed using both Pearson and Spearman methods, and the results are compiled into
a comprehensive data frame. Parallel processing is supported to expedite computations.
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
