% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalCCI2.R
\name{TemporalCCI2}
\alias{TemporalCCI2}
\title{Temporal Cell-Cell Interaction Analysis with Discrete Pseudotime Binning}
\usage{
TemporalCCI2(
  seu.obj,
  Pseudotime.name = "Pseudotime",
  CellType.name = "Cell.Type",
  discrete.numer = 20,
  CCI.database = CellChat::CellChatDB.human,
  interaction.name = "interaction_name_2"
)
}
\arguments{
\item{seu.obj}{A Seurat object containing single-cell RNA-seq data and metadata.}

\item{Pseudotime.name}{A string specifying the name of the pseudotime annotation in the Seurat object metadata. Default is "Pseudotime".}

\item{CellType.name}{A string specifying the name of the cell type annotation in the Seurat object metadata. Default is "Cell.Type".}

\item{discrete.numer}{An integer specifying the number of discrete pseudotime intervals to create. Default is 20.}

\item{CCI.database}{A CellChat database object containing ligand-receptor interactions. Default is \code{CellChat::CellChatDB.human}.}

\item{interaction.name}{A string specifying the interaction name column in the CCI database. Default is "interaction_name_2".}
}
\value{
A list containing:
\item{cor.res}{A data frame with correlation results between interaction probabilities and pseudotime stages, including Pearson and Spearman correlation coefficients and p-values.}
\item{cal.res.wide}{A wide-format data frame with interaction probabilities across pseudotime stages.}
\item{cal.res.long}{A long-format data frame with interaction probabilities and corresponding pseudotime stages.}
\item{cellchat}{The CellChat object used for the analysis, containing all intermediate results.}
}
\description{
This function performs cell-cell interaction (CCI) analysis over discrete pseudotime intervals.
It utilizes the CellChat framework to identify and analyze ligand-receptor interactions across different
stages of pseudotime, and computes correlations between interaction probabilities and pseudotime stages.
}
\details{
This function first divides the pseudotime into discrete intervals and assigns each cell to an interval. It then uses
the CellChat package to analyze cell-cell communication based on ligand-receptor interactions, computing the communication
probabilities for each interaction across the pseudotime intervals. The function also performs correlation analyses
between these probabilities and the pseudotime stages, providing insights into how interactions evolve over time.
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
